move all functions to main.py (in single file)
